requirements.txt
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-mysql==5.3.1
pandas>=2.2.0
requests>=2.28.2
